refs #75 split Qorus debug stuff
